Module EDSL.Elist

Functions on 'a list t values.

val make : 'a t list -> 'a list t

Make an EDSL list out of an OCaml list.

val append : 'a list t -> 'a list t -> 'a list t

Concatenate two EDSL lists.

val iter : 'a list t -> f:((unit -> 'a t) -> unit t) -> unit t

Iterate over a list, the body of the loop ~f takes as argument function that returns the current eletment at the EDSL level.

val serialize_byte_array_list : str list t -> str t

val deserialize_to_byte_array_list : str t -> str list t

val serialize_str_list : str list t -> str t

val deserialize_to_str_list : str t -> str list t

val serialize_int_list : int list t -> str t

val deserialize_to_int_list : str t -> int list t