Module Transform.Constant_propagation

A basic implementation of constant propagation.

val process : ?⁠trace:Stdlib.Format.formatter -> 'a Language.t -> 'a Language.t

Simplify an 'a EDSL.t expression by propagating some of the constant values.

val test : unit -> unit

Some tests specific to the module, see option "--run-constant-propagation-tests" of the main tests.